+++ README.md
@@ -14,12 +14,16 @@ Library functionality is split over the following classes:
 
 * `PGS_CirclePacking`
   * Circle packings of shapes, subject to varying constraints and patterns of tangencies
+* `PGS_Coloring`
+  * TODO
 * `PGS_Construction`
   * Construct uncommon/interesting 2D primitives
 * `PGS_Contour`
   * Methods that produce various contours from shapes: medial axes, straight skeletons, offset curves, etc.
 * `PGS_Conversion`
-  * Conversion between PShapes and JTS Geometries 
+  * Conversion between *Processing* PShapes and *JTS* Geometries 
+* `PGS_Meshing`
+  * Mesh generation (excluding triangulation).
 * `PGS_Morphology`
   * Methods that affect the geometry or topology of shapes (buffering, simplification, smoothing, etc.)
 * `PGS_Optimsation`
@@ -362,7 +366,7 @@ Much of the functionality (but by no means all) is demonstrated below:
   </tr>
 </table>
 
-## Triangulation
+## *Triangulation*
 
 <table>
   <tr>
@@ -378,27 +382,40 @@ Much of the functionality (but by no means all) is demonstrated below:
 
   <tr>
     <td align="center" valign="center"><b>Poisson Delaunay Triangulation</td>
-    <td align="center" valign="center" colspan="2"><b>Urquhart Faces</td>
-    <td align="center" valign="center"><b>Gabriel Faces</td>
   </tr>
   <tr>
     <td valign="top" width="25%"><img src="resources/contour/poissonTriangulation.gif"></td>
-    <td valign="top" width="25%"><img src="resources/triangulation/urquhart1.png"></td>
-    <td valign="top" width="25%"><img src="resources/triangulation/urquhart2.png"></td>
-    <td valign="top" width="25%"><img src="resources/triangulation/gabriel1.png"></td>
   </tr>
 
   <tr>
     <td align="center" valign="center"><i>Delaunay triangulation</i> of shapes where <i>steiner points</i> generated by poisson disk sampling are inserted.</td>
+  </tr>
+</table>
+
+## *Meshing*
+
+<table>
+  <tr>
+    <td align="center" valign="center" colspan="2"><b>Urquhart Faces</td>
+    <td align="center" valign="center"><b>Gabriel Faces</td>
+    <td align="center" valign="center"><b>Triangulation Dual</td>
+  </tr>
+  <tr>
+    <td valign="top" width="25%"><img src="resources/triangulation/urquhart1.png"></td>
+    <td valign="top" width="25%"><img src="resources/triangulation/urquhart2.png"></td>
+    <td valign="top" width="25%"><img src="resources/triangulation/gabriel1.png"></td>
+    <td valign="top" width="25%"><img src="resources/triangulation/triangulationDual.png"></td>
+  </tr>
+  <tr>
     <td align="center" valign="center" colspan="2">Polygon faces of an <i>Urquhart Graph</i> (derived from a triangulation).</td>
     <td align="center" valign="center">Polygon faces of a <i>Gabriel Graph</i> (derived from a triangulation).</td>
   </tr>
 
   <tr>
-    <td align="center" valign="center"><b>Triangulation Dual</td>
+    <td align="center" valign="center"><b>Quadrangulation</td>
   </tr>
   <tr>
-    <td valign="top" width="25%"><img src="resources/triangulation/triangulationDual.png"></td>
+    <td valign="top" width="25%"><img src="resources/triangulation/quadrangulation.png"></td>
   </tr>
 </table>
