+++ src/main/java/micycle/pgs/PGS_Triangulation.java
@@ -1,11 +1,9 @@
 package micycle.pgs;
 
 import static micycle.pgs.PGS_Conversion.fromPShape;
-import static micycle.pgs.PGS_Conversion.toPShape;
 
 import java.util.ArrayList;
 import java.util.Collection;
-import java.util.HashSet;
 import java.util.List;
 import java.util.function.Consumer;
 
@@ -16,24 +14,17 @@ import org.locationtech.jts.geom.Envelope;
 import org.locationtech.jts.geom.Geometry;
 import org.locationtech.jts.geom.LinearRing;
 import org.locationtech.jts.geom.Location;
-import org.locationtech.jts.geom.Polygon;
-import org.locationtech.jts.operation.polygonize.Polygonizer;
-import org.locationtech.jts.operation.polygonize.QuickPolygonizer;
 import org.tinfour.common.IConstraint;
-import org.tinfour.common.IIncrementalTin;
-import org.tinfour.common.IQuadEdge;
 import org.tinfour.common.PolygonConstraint;
 import org.tinfour.common.SimpleTriangle;
 import org.tinfour.common.Vertex;
 import org.tinfour.standard.IncrementalTin;
 import org.tinfour.utils.TriangleCollector;
-import org.tinspin.index.PointIndex;
-import org.tinspin.index.kdtree.KDTree;
+
 import earcut4j.Earcut;
 import micycle.pgs.PGS.LinearRingIterator;
 import micycle.pgs.color.RGB;
-import micycle.pgs.utility.IncrementalTinDual;
-import micycle.pgs.utility.PoissonDistribution;
+import micycle.pgs.utility.Nullable;
 import processing.core.PConstants;
 import processing.core.PShape;
 import processing.core.PVector;
@@ -91,8 +82,8 @@ public final class PGS_Triangulation {
 	 * @see #delaunayTriangulationPoints(PShape, List, boolean, int, boolean)
 	 * @see #delaunayTriangulationMesh(PShape, List, boolean, int, boolean)
 	 */
-	public static PShape delaunayTriangulation(PShape shape, Collection<PVector> steinerPoints, boolean constrain, int refinements,
-			boolean pretty) {
+	public static PShape delaunayTriangulation(PShape shape, @Nullable Collection<PVector> steinerPoints, boolean constrain,
+			int refinements, boolean pretty) {
 		final IncrementalTin tin = delaunayTriangulationMesh(shape, steinerPoints, constrain, refinements, pretty);
 
 		final PShape triangulation = new PShape(PConstants.GROUP);
@@ -179,7 +170,7 @@ public final class PGS_Triangulation {
 	 * @see #delaunayTriangulationPoints(PShape, List, boolean, int, boolean)
 	 * @see #delaunayTriangulationMesh(PShape, List, boolean, int, boolean)
 	 */
-	public static List<PVector> delaunayTriangulationPoints(PShape shape, Collection<PVector> steinerPoints, boolean constrain,
+	public static List<PVector> delaunayTriangulationPoints(PShape shape, @Nullable Collection<PVector> steinerPoints, boolean constrain,
 			int refinements, boolean pretty) {
 		final IncrementalTin tin = delaunayTriangulationMesh(shape, steinerPoints, constrain, refinements, pretty);
 
@@ -261,7 +252,7 @@ public final class PGS_Triangulation {
 	 * @see #delaunayTriangulation(PShape, List, boolean, int, boolean)
 	 * @see #delaunayTriangulationPoints(PShape, List, boolean, int, boolean)
 	 */
-	public static IncrementalTin delaunayTriangulationMesh(PShape shape, Collection<PVector> steinerPoints, boolean constrain,
+	public static IncrementalTin delaunayTriangulationMesh(PShape shape, @Nullable Collection<PVector> steinerPoints, boolean constrain,
 			int refinements, boolean pretty) {
 		Geometry g = shape == null ? PGS.GEOM_FACTORY.createEmpty(2) : fromPShape(shape);
 		final IncrementalTin tin = new IncrementalTin(10);
@@ -470,173 +461,7 @@ public final class PGS_Triangulation {
 		return triangulation;
 	}
 
-	/**
-	 * Generates a shape consisting of polygonal faces of an <i>Urquhart graph</i>.
-	 * An Urquhart graph is obtained by removing the longest edge from each triangle
-	 * in a triangulation.
-	 * <p>
-	 * In practice this is a way to tessellate a shape into polygons (with the
-	 * resulting tessellation being in between a
-	 * {@link #delaunayTriangulation(PShape, List, boolean, int, boolean)
-	 * triangulation} and a {@link micycle.pgs.PGS_Processing#partition(PShape)
-	 * partition}).
-	 * <p>
-	 * Note that this method processes a Delaunay triangulation. Process a shape
-	 * using
-	 * {@link #delaunayTriangulationMesh(PShape, Collection, boolean, int, boolean)
-	 * delaunayTriangulationMesh()} first and then feed it to this method.
-	 * 
-	 * @param triangulation     a triangulation mesh
-	 * @param preservePerimeter whether to preserve the perimeter of the input
-	 *                          triangulation; when true, retains edges that lie on
-	 *                          the perimeter of the triangulation mesh that would
-	 *                          have otherwise been removed according to the
-	 *                          urquhart condition.
-	 * @return a GROUP PShape where each child shape is a single face
-	 * @since 1.1.0
-	 * @see #gabrielFaces(IncrementalTin)
-	 */
-	@SuppressWarnings("unchecked")
-	public static PShape urquhartFaces(final IIncrementalTin triangulation, final boolean preservePerimeter) {
-		final HashSet<IQuadEdge> edges = new HashSet<>();
-		final HashSet<IQuadEdge> uniqueLongestEdges = new HashSet<>();
-
-		final boolean notConstrained = triangulation.getConstraints().isEmpty();
-
-		TriangleCollector.visitSimpleTriangles(triangulation, t -> {
-			final IConstraint constraint = t.getContainingRegion();
-			if (notConstrained || (constraint != null && constraint.definesConstrainedRegion())) {
-				edges.add(t.getEdgeA().getBaseReference());
-				edges.add(t.getEdgeB().getBaseReference());
-				edges.add(t.getEdgeC().getBaseReference());
-				final IQuadEdge longestEdge = findLongestEdge(t).getBaseReference();
-				if (!preservePerimeter || (preservePerimeter && !longestEdge.isConstrainedRegionBorder())) {
-					uniqueLongestEdges.add(longestEdge);
-				}
-			}
-		});
-
-		final Polygonizer polygonizer = new QuickPolygonizer(false);
-		polygonizer.setCheckRingsValid(false);
-		edges.removeAll(uniqueLongestEdges);
-		edges.forEach(edge -> polygonizer
-				.add(PGS.GEOM_FACTORY.createLineString(new Coordinate[] { toCoord(edge.getA()), toCoord(edge.getB()) })));
-
-		final PShape out = new PShape(PConstants.GROUP);
-		polygonizer.getPolygons().forEach(p -> {
-			final PShape face = toPShape((Polygon) p);
-			face.setStrokeWeight(3);
-			out.addChild(face);
-		});
-
-		return out;
-	}
-
-	/**
-	 * Generates a shape consisting of polygonal faces of a <i>Gabriel graph</i>. A
-	 * Gabriel graph is obtained by removing each edge E from a triangulation if a
-	 * vertex lies within a circle of diameter = length(E), centered on the midpoint
-	 * of E.
-	 * <p>
-	 * In practice this is a way to tessellate a shape into polygons (with the
-	 * resulting tessellation being reminiscent of shattering the shape as if it
-	 * were glass).
-	 * <p>
-	 * Note that this method processes a Delaunay triangulation. Process a shape
-	 * using
-	 * {@link #delaunayTriangulationMesh(PShape, Collection, boolean, int, boolean)
-	 * delaunayTriangulationMesh()} first and then feed it to this method.
-	 * 
-	 * @param triangulation     a triangulation mesh
-	 * @param preservePerimeter whether to preserve the perimeter of the input
-	 *                          triangulation; when true, retains edges that lie on
-	 *                          the perimeter of the triangulation mesh that would
-	 *                          have otherwise been removed according to the
-	 *                          urquhart condition.
-	 * @return a GROUP PShape where each child shape is a single face
-	 * @since 1.1.0
-	 * @see #urquhartFaces(IncrementalTin, boolean)
-	 */
-	@SuppressWarnings("unchecked")
-	public static PShape gabrielFaces(final IIncrementalTin triangulation, final boolean preservePerimeter) {
-		final HashSet<IQuadEdge> edges = new HashSet<>();
-		final HashSet<Vertex> vertices = new HashSet<>();
-
-		final boolean notConstrained = triangulation.getConstraints().isEmpty();
-		TriangleCollector.visitSimpleTriangles(triangulation, t -> {
-			final IConstraint constraint = t.getContainingRegion();
-			if (notConstrained || (constraint != null && constraint.definesConstrainedRegion())) {
-				edges.add(t.getEdgeA().getBaseReference()); // add edge to set
-				edges.add(t.getEdgeB().getBaseReference()); // add edge to set
-				edges.add(t.getEdgeC().getBaseReference()); // add edge to set
-				vertices.add(t.getVertexA());
-				vertices.add(t.getVertexB());
-				vertices.add(t.getVertexC());
-			}
-		});
-
-		final PointIndex<Vertex> tree = KDTree.create(2, (p1, p2) -> {
-			final double deltaX = p1[0] - p2[0];
-			final double deltaY = p1[1] - p2[1];
-			return Math.sqrt(deltaX * deltaX + deltaY * deltaY);
-		});
-		vertices.forEach(v -> tree.insert(new double[] { v.x, v.y }, v));
-
-		final HashSet<IQuadEdge> nonGabrielEdges = new HashSet<>(); // base references to edges that should be removed
-		edges.forEach(edge -> {
-			final double[] midpoint = midpoint(edge);
-			final Vertex near = tree.query1NN(midpoint).value();
-			if (near != edge.getA() && near != edge.getB()) {
-				if (!preservePerimeter || (preservePerimeter && !edge.isConstrainedRegionBorder())) {
-					nonGabrielEdges.add(edge); // base reference
-				}
-			}
-		});
-		edges.removeAll(nonGabrielEdges);
-
-		final Polygonizer polygonizer = new QuickPolygonizer(false);
-		polygonizer.setCheckRingsValid(false);
-		edges.forEach(edge -> polygonizer
-				.add(PGS.GEOM_FACTORY.createLineString(new Coordinate[] { toCoord(edge.getA()), toCoord(edge.getB()) })));
-
-		final PShape out = new PShape(PConstants.GROUP);
-		polygonizer.getPolygons().forEach(p -> {
-			final PShape face = toPShape((Polygon) p);
-			face.setStrokeWeight(3);
-			out.addChild(face);
-		});
-		return out;
-	}
-
-	/**
-	 * Generates a shape consisting of polygonal faces of the dual graph of the
-	 * given triangulation.
-	 * <p>
-	 * In practice, the resulting dual mesh has hexagonal-like cells.
-	 * 
-	 * <p>
-	 * If the input has been generated from a PShape, consider generating the
-	 * triangulation with refinements > 1 for better dual mesh results.
-	 * 
-	 * @param triangulation a triangulation mesh
-	 * @return a GROUP PShape where each child shape is a single face
-	 * @since 1.2.0
-	 */
-	public static PShape dualFaces(final IIncrementalTin triangulation) {
-		final IncrementalTinDual dual = new IncrementalTinDual(triangulation);
-		final PShape dualMesh = dual.getMesh();
-		PGS_Conversion.setAllFillColor(dualMesh, RGB.WHITE);
-		PGS_Conversion.setAllStrokeColor(dualMesh, RGB.PINK, 2);
-		return dualMesh;
-	}
-
-	private static double[] midpoint(final IQuadEdge edge) {
-		final Vertex a = edge.getA();
-		final Vertex b = edge.getB();
-		return new double[] { (a.x + b.x) / 2d, (a.y + b.y) / 2d };
-	}
-
-	private static PVector toPVector(final Vertex v) {
+	static PVector toPVector(final Vertex v) {
 		return new PVector((float) v.getX(), (float) v.getY());
 	}
 
@@ -652,28 +477,5 @@ public final class PGS_Triangulation {
 		double y = a.y + b.y + c.y;
 		y /= 3;
 		return new Coordinate(x, y);
-	}
-
-	private static Coordinate toCoord(final Vertex v) {
-		return new Coordinate(v.x, v.y);
-	}
-
-	/**
-	 * Calculate the longest edge of a given triangle.
-	 */
-	private static IQuadEdge findLongestEdge(final SimpleTriangle t) {
-		if (t.getEdgeA().getLength() > t.getEdgeB().getLength()) {
-			if (t.getEdgeC().getLength() > t.getEdgeA().getLength()) {
-				return t.getEdgeC();
-			} else {
-				return t.getEdgeA();
-			}
-		} else {
-			if (t.getEdgeC().getLength() > t.getEdgeB().getLength()) {
-				return t.getEdgeC();
-			} else {
-				return t.getEdgeB();
-			}
-		}
 	}
 }
